A lot of progress in getting multi-team pitchers working in the frontend, but there's still an issue to solve in figuring out why the final data isn't being sent from the endpoint.

+++ config/ratings_guide_db(schema).sql
@@ -134,7 +134,7 @@ CREATE TABLE `multi_team_pitchers` (
     FOREIGN KEY (`real_team_id`) REFERENCES `real_teams`(`real_team_id`) ON DELETE RESTRICT ON UPDATE CASCADE,
     `pitcher` varchar(30) NOT NULL,
     `throws` varchar(1) NOT NULL,
-    `ip` int unsigned NOT NULL
+    `ip` decimal(5,2) NOT NULL
 ) ENGINE=InnoDB  DEFAULT CHARSET=utf8mb4;
 
 -- --------------------------------------------------------
